+++ WhyNComing/src/main/java/org/sparta/whyncoming/order/domain/entity/Cart.java
@@ -1,0 +1,77 @@
+package org.sparta.whyncoming.order.domain.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.sparta.whyncoming.product.domain.entity.Product;
+import org.sparta.whyncoming.store.domain.entity.Store;
+import org.sparta.whyncoming.user.domain.entity.User;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+@Entity
+@Table(name = "carts")
+@Getter
+@NoArgsConstructor
+@EntityListeners(AuditingEntityListener.class)
+public class Cart {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID cartId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "store_id", nullable = false)
+    private Store store;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "product_id", nullable = false)
+    private Product product;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "order_id", nullable = true) // ERD 상 Order가 저장되어 있지 않아도 가능한 것으로 봄
+    private Order order;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_no", nullable = false)
+    private User user;
+
+    @Column(nullable = false)
+    private Integer quantity = 1;
+
+    @CreatedDate
+    @Column(nullable = false, updatable = false)
+    private LocalDateTime createdDate;
+
+    @LastModifiedDate
+    @Column(nullable = false)
+    private LocalDateTime modifiedDate;
+
+    @Column
+    private LocalDateTime deletedDate;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "created_by")
+    private User createdBy;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "modified_by")
+    private User modifiedBy;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "deleted_by")
+    private User deletedBy;
+
+
+    public Cart(Store store, Product product, Order order, User user, Integer quantity) {
+        this.store = store;
+        this.product = product;
+        this.order = order;
+        this.user = user;
+        this.quantity = quantity;
+    }
+}
