+++ WhyNComing/src/main/java/org/sparta/whyncoming/order/domain/entity/Order.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.sparta.whyncoming.order.domain.enums.Status;
-import org.sparta.whyncoming.product.domain.entity.Cart;
 import org.sparta.whyncoming.store.domain.entity.Store;
 import org.sparta.whyncoming.user.domain.entity.User;
 import org.springframework.data.annotation.CreatedDate;
