+++ WhyNComing/src/main/java/org/sparta/whyncoming/user/domain/entity/User.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 import org.sparta.whyncoming.order.domain.entity.Order;
 import org.sparta.whyncoming.order.domain.entity.OwnerReview;
 import org.sparta.whyncoming.order.domain.entity.Review;
-import org.sparta.whyncoming.product.domain.entity.Cart;
+import org.sparta.whyncoming.order.domain.entity.Cart;
 import org.sparta.whyncoming.store.domain.entity.Store;
 import org.sparta.whyncoming.user.domain.enums.Role;
 import org.springframework.data.annotation.CreatedDate;
